Improved logging p1
