+++ Cargo.toml
@@ -8,8 +8,6 @@ edition = "2021"
 members = ["crates/*"]
 
 [dependencies]
-chrono = "0.4.19"
-image = "0.23.14"
 rayon = "1.5.1"
 rand = { version = "0.8.3", features = [ "small_rng" ] }
 rand_seeder = "0.2.2"
@@ -21,9 +19,10 @@ wavefront_obj = "10.0.0"
 winit = { version = "0.26.1", optional = true }
 region = { path = "./crates/region" }
 loader = { path = "./crates/loader" }
-simple_logger = "4.0"
+output = { path = "./crates/output" }
 clap = { version = "4.1.8", features = ["derive", "wrap_help"] }
 gui = { path = "./crates/gui" }
+indicatif = "0.17.3"
 
 [features]
 f64 = ["implementations/f64"]
