+++ crates/gui/Cargo.toml
@@ -9,5 +9,4 @@ vulkano-shaders = { version = "0.28.0" }
 vulkano-win = { version = "0.28.0" }
 winit = { version = "0.26.1" }
 implementations = { path = "../implementations" }
-chrono = "0.4.19"
-image = "0.23.14"
+indicatif = "0.17.3"
